feat: 채팅 기능 추가(채팅방 등록+조회, 채팅메세지 등록+조회) [S12P11D205-215]

+++ backend/igeolu/src/main/java/com/ssafy/igeolu/domain/chatmessage/entity/UserRoomStatus.java
@@ -17,8 +17,8 @@ import lombok.NoArgsConstructor;
 public class UserRoomStatus {
 	@Id
 	private ObjectId id;
-	private Long userId;
-	private Long roomId;
+	private Integer userId;
+	private Integer roomId;
 
 	// 유저가 해당방에서 마지막으로 읽은 메시지 ID
 	private ObjectId lastReadMessageId;
