+++ backend/igeolu/src/main/java/com/ssafy/igeolu/domain/chatmessage/repository/UserRoomStatusRepository.java
@@ -8,5 +8,5 @@ import reactor.core.publisher.Mono;
 
 public interface UserRoomStatusRepository extends ReactiveMongoRepository<UserRoomStatus, String> {
 	// 특정 사용자의 읽음 상태 조회
-	Mono<UserRoomStatus> findByUserIdAndRoomId(Long userId, Long roomId);
+	Mono<UserRoomStatus> findByUserIdAndRoomId(Integer userId, Integer roomId);
 }
