+++ backend/igeolu/src/main/java/com/ssafy/igeolu/facade/chatmessage/dto/request/ChatMessagePostRequestDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ChatMessagePostRequestDto {
 
-	private Long roomId;
-	private Long writerId;
+	private Integer roomId;
+	private Integer writerId;
 	private String content;
 }
