+++ backend/igeolu/src/main/java/com/ssafy/igeolu/facade/chatroom/dto/request/ChatRoomPostRequestDto.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ChatRoomPostRequestDto {
 
-	private Long memberId;
-	private Long realtorId;
+	private Integer memberId;
+	private Integer realtorId;
 }
